Rename test annotator to instance annotator
