+++ client/annotation_client.py
@@ -8,7 +8,7 @@ from kivy.uix.screenmanager import ScreenManager
 
 from client.client_config import ClientConfig
 from client.screens.image_view_screen import ImageViewScreen
-from client.screens.test_annotator import InstanceAnnotatorScreen
+from client.screens.instance_annotator_screen import InstanceAnnotatorScreen
 from client.screens.project_select_screen import ProjectSelectScreen
 from client.screens.project_tool_screen import ProjectToolScreen
